Move analysis table to the left panel analysis folder
